refactor: PrezelTabs 컴포넌트 구조 개선 및 패키지 이동
- 탭 컴포넌트의 재사용성을 높이기 위해 레이아웃 로직을 분리하고, 관련 파일을 `navigations` 패키지로 이동 및 정리했습니다.
