+++ Prezel/core/designsystem/src/main/java/com/team/prezel/core/designsystem/component/navigations/PrezelTabSize.kt
@@ -1,0 +1,3 @@
+package com.team.prezel.core.designsystem.component.navigations
+
+enum class PrezelTabSize { REGULAR, MEDIUM }
